+++ packages/ui/src/collapsible-card/stories/index.story.tsx
@@ -28,6 +28,7 @@ function Text( { children }: { children: React.ReactNode } ) {
 const meta: Meta< typeof CollapsibleCard.Root > = {
 	tags: [ 'manifest' ],
 	title: 'Design System/Components/CollapsibleCard',
+	synonyms: [ 'accordion', 'card', 'expand' ],
 	component: CollapsibleCard.Root,
 	subcomponents: {
 		'CollapsibleCard.Header': CollapsibleCard.Header,