+++ packages/ui/src/collapsible/stories/index.story.tsx
@@ -5,6 +5,7 @@ import * as Collapsible from '../index';
 const meta: Meta< typeof Collapsible.Root > = {
 	tags: [ 'manifest' ],
 	title: 'Design System/Components/Collapsible',
+	synonyms: [ 'accordion', 'disclosure', 'expand' ],
 	component: Collapsible.Root,
 	subcomponents: {
 		'Collapsible.Trigger': Collapsible.Trigger,